revised anzctr dates
Manually checked and updated the completion date for anything in which the current last_updated_date does not match the archived one from early January (meaning something has been updated since our Phase 2 scrape). Using the ANZCTR history page, I was able to cross-check the date as it stood in early January.

Also updated the ANZCTR scraper with the fixed code to get the correct dates.