Fixed for Phase 3 ANZCTR
Took updated scrape data and, as in Phase 2, looked for everything that had been updated since our Phase 3 scrape in Early July 2021. I think cross-referenced it with the history page on the ANZCTR in order to derive the completion date as it stood in early July. The appropriate datasets in the repo were updated accordingly.